backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
PyJWT==2.9.0
cryptography==43.0.1
supabase==2.9.0
httpx==0.27.2
anthropic==0.34.2
python-dotenv==1.0.1
